backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
umap-learn==0.5.6
scipy==1.13.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
